feat: move mock helper to src test and rename to mock-data
